Always convert truthy booleans to 1
This changes `c10::convert` to interpret booleans as a raw byte and
convert to the integer 1 or 0 before converting to the final type.
This requires some careful work as the compiler want to optimize with
the assumption that a `bool` will always be 1 or 0.

The added test now passes for most operators which support bools via
conversion, but fails for custom kernels.

ghstack-source-id: 527796fd260b61e35957a2396fe95ea66bcff069
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77122